Transfer controls from View to ViewModel
