--- VocabularyManager/Models/SystemVolumeConfigurator.cs
@@ -1,40 +1,0 @@
-﻿using NAudio.CoreAudioApi;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-
-namespace VocabularyManager.Models
-{
-    public class SystemVolumeConfigurator
-    {
-        private readonly MMDeviceEnumerator _deviceEnumerator = new MMDeviceEnumerator();
-        private readonly MMDevice _playbackDevice;
-
-        public SystemVolumeConfigurator()
-        {
-            _playbackDevice = _deviceEnumerator.GetDefaultAudioEndpoint(DataFlow.Render, Role.Multimedia);
-            _playbackDevice.AudioEndpointVolume.OnVolumeNotification += AudioEndpointVolume_OnVolumeNotification;
-        }
-
-        void AudioEndpointVolume_OnVolumeNotification(AudioVolumeNotificationData data)
-        {
-            MessageBox.Show((data.MasterVolume*100).ToString());
-        }
-
-        public float GetVolume()
-        {
-            return (float)(_playbackDevice.AudioEndpointVolume.MasterVolumeLevelScalar * 100);
-        }
-
-        public void SetVolume(float volumeLevel)
-        {
-            if (volumeLevel < 0 || volumeLevel > 100)
-                throw new ArgumentException("Volume must be between 0 and 100!");
-
-            _playbackDevice.AudioEndpointVolume.MasterVolumeLevelScalar = volumeLevel / 100.0f;
-        }
-    }
-}
