+++ VocabularyManager/ViewModel/MainViewModel.cs
@@ -1,9 +1,11 @@
 using System;
 using System.Collections.ObjectModel;
+using System.Windows.Controls;
 using GalaSoft.MvvmLight;
 using GalaSoft.MvvmLight.Command;
 using NamedPipeWrapper;
 using VocabularyManager.Models;
+using VocabularyManagerService.Services;
 
 namespace VocabularyManager.ViewModel
 {
@@ -11,7 +13,11 @@ namespace VocabularyManager.ViewModel
     public class MainViewModel : ViewModelBase
     {
         private ObservableCollection<VolumeModel> volumeModels;
-        private readonly NamedPipeClient<string> _client = new NamedPipeClient<string>("test");
+        private readonly NamedPipeClient<string> _client = new NamedPipeClient<string>("nana");
+        private VolumeModel volumeModel;
+        private float currentValue;
+        private string mainTb;
+
         public ObservableCollection<VolumeModel> Volume
         {
             get { return volumeModels; }
@@ -22,8 +28,6 @@ namespace VocabularyManager.ViewModel
             }
         }
 
-        private VolumeModel volumeModel;
-
         public VolumeModel VolumeModel
         {
             get { return volumeModel; }
@@ -33,8 +37,6 @@ namespace VocabularyManager.ViewModel
                 RaisePropertyChanged("Volume");
             }
         }
-
-        private float currentValue;
 
         public float CurrentValue
         {
@@ -50,11 +52,23 @@ namespace VocabularyManager.ViewModel
 
         public RelayCommand GetVolumeCommand { get; set; }
 
+        public string SystemValue { get; set; }
+
+        public string MainTb
+        {
+            get { return this.mainTb; }
+            set
+            {
+                mainTb = value;
+                RaisePropertyChanged("MainTb");
+            }
+        }
+
         public MainViewModel()
         {
-            //_client.ServerMessage += OnServerMessage;
-            //_client.Disconnected += OnDisconnected;
-            //_client.Start();
+            _client.ServerMessage += OnServerMessage;
+            _client.Disconnected += OnDisconnected;
+            _client.Start();
 
             VolumeModel = new VolumeModel();
             Volume = new ObservableCollection<VolumeModel>();
@@ -62,52 +76,30 @@ namespace VocabularyManager.ViewModel
             GetVolumeCommand = new RelayCommand(GetVolume);
         }
 
-        void Apply()
+        private void Apply()
         {
             _client.PushMessage(currentValue.ToString());
         }
 
-        public string SystemValue { get; set; }
-
-        public string RichTb { get; set; }
-
-        void GetVolume()
+        private void GetVolume()
         {
-            SystemVolumeConfigurator volume = new SystemVolumeConfigurator();
+            SystemVolumeService volume = new SystemVolumeService();
 
             CurrentValue = volume.GetVolume();
 
             SystemValue = CurrentValue.ToString();
+
+            _client.PushMessage(SystemValue);
         }
 
-        //private void OnServerMessage(NamedPipeConnection<string, string> connection, string message)
-        //{
-        //    //richTextBox.Dispatcher.Invoke(new Action(delegate
-        //    //{
-        //    //    AddLine("Server: " + message);
-        //    //}));
+        private void OnServerMessage(NamedPipeConnection<string, string> connection, string message)
+        {
+            MainTb += Environment.NewLine + "Server" + message;
+        }
 
-        //    RichTb += "Server: " + message;
-        //}
-
-        //private void OnDisconnected(NamedPipeConnection<string, string> connection)
-        //{
-        //    //richTextBox.Dispatcher.Invoke(new Action(delegate
-        //    //{
-        //    //    AddLine("Disconnected from server");
-        //    //}));
-        //    RichTb += "Disconnected from server";
-        //}
-
-        //private void AddLine(string html)
-        //{
-        //    //richTextBox.Dispatcher.Invoke(new Action(delegate
-        //    //{
-        //    //    richTextBox.AppendText(Environment.NewLine + html);
-        //    //}));
-
-        //    RichTb += Environment.NewLine + html;
-        //}
-
+        private void OnDisconnected(NamedPipeConnection<string, string> connection)
+        {
+            MainTb += Environment.NewLine + "Disconnected from server";
+        }
     }
 }
