+++ VocabularyManager/Views/MainWindow.xaml
@@ -20,17 +20,16 @@
             <ColumnDefinition></ColumnDefinition>
         </Grid.ColumnDefinitions>
         <TextBox x:Name="textBox" Grid.Row="1" Grid.Column="1" Margin="77.8,170.6,87.2,4.4" IsReadOnly="True" Text ="{Binding SystemValue}"/>
-        <Button x:Name="button" Content="Apply" Margin="9.8,70,39.6,0" Grid.Column="2"  Height="30" VerticalAlignment="Top" Command="{Binding ApplyCommand}" CommandParameter="{Binding CurrentValue}" Click="button_Click"/>
+        <Button x:Name="button" Content="Apply" Margin="9.8,70,39.6,0" Grid.Column="2"  Height="30" VerticalAlignment="Top" Command="{Binding ApplyCommand}" CommandParameter="{Binding CurrentValue}"/>
         <Label x:Name="label" Content="Audio" HorizontalAlignment="Left" Grid.Column="1" Margin="0,50,0,200" Grid.RowSpan="2" />
-        <Slider Value="{Binding CurrentValue}" Minimum="0" Maximum="100" SelectionStart="0" x:Name="slider" Grid.Column="1" Grid.Row="0" Margin="0,80,150,200" ValueChanged="slider_ValueChanged" Grid.ColumnSpan="2" Grid.RowSpan="2" />
+        <Slider Value="{Binding CurrentValue}" Minimum="0" Maximum="100" SelectionStart="0" x:Name="slider" Grid.Column="1" Grid.Row="0" Margin="0,80,150,200"  Grid.ColumnSpan="2" Grid.RowSpan="2" />
         <Button x:Name="btnGet" Content="Get" Grid.Row="2" Grid.Column="1" Margin="50,30,50,50" Command="{Binding GetVolumeCommand}"/>
-        <RichTextBox x:Name="richTextBox" HorizontalAlignment="Left" Height="149" Margin="46.8,7.6,0,0" Grid.Row="1" VerticalAlignment="Top" Width="191" Grid.Column="1" HorizontalScrollBarVisibility="Visible" VerticalScrollBarVisibility="Visible">
+        <RichTextBox HorizontalAlignment="Left" Height="149" Margin="46.8,7.6,0,0" Grid.Row="1" VerticalAlignment="Top" Width="191" Grid.Column="1" HorizontalScrollBarVisibility="Visible" VerticalScrollBarVisibility="Visible">
             <FlowDocument>
                 <Paragraph>
-                    <Run Text="{}"/>
+                    <Run Text="{Binding Path=MainTb}"/>
                 </Paragraph>
             </FlowDocument>
         </RichTextBox>
-
     </Grid>
 </Window>
