+++ VocabularyManager/VocabularyManager.csproj
@@ -88,7 +88,6 @@
       <SubType>Designer</SubType>
     </ApplicationDefinition>
     <Compile Include="Interfaces\IDataAccessService.cs" />
-    <Compile Include="Models\SystemVolumeConfigurator.cs" />
     <Page Include="Views\MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
